requirements.txt
aiogram==2.15
python-dotenv==1.0.1